SHRINKDESC-24 - Problems with the version attribute in the peristence element
The version attribute wasn't appearing because it was getting it from the field which didn't have a default value . 
The only way to make it appear is to set it manually which couldn't be done because it wasn't exposed in the model. 
I've set the version to default to 2.0 and exposed the version in the model. I've also added two test cases to ensure that the version appears by default and that it can be set from code.