SHRINKDESC-20  - Handle importing from empty input streams
This fix enables the developer to import a descriptor from a stream that might be empty by returning a default, blank instance of the descriptor if it is found to be empty (instead of throwing an exception as it does now).
It does this by converting the stream to a string and then always using the from(String string) method to load the descriptor model. This method includes the check for an empty string and if the string content is empty, then an instance of the view class is requested and returned. We pass the user view class into the constructor for the SchemaDescriptorImporter.